backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Bcrypt==1.0.1
psycopg2-binary==2.9.7
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.0.1
marshmallow==3.20.1
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
gunicorn==21.2.0
python-dateutil==2.8.2
geopy==2.4.0 